+++ src/grid/GridBody.tsx
@@ -3,6 +3,12 @@ import * as React from 'react';
 export type BodyCellRender = (rowIndex: number, columnIndex: number) => React.ReactNode;
 
 export interface GridBodyProps extends React.DivProps {
+    rowCount: number;
+    /**
+     * Body row height.  
+     * To adjust the head height set the head `style` prop.
+     */
+    rowHeight: number | SizeCallback;
     children?: BodyCellRender;
 }